Removed `message.go` file contained references to whisper message and envelope code. All that whisper code has been copied and distilled into swarm/pss/crypto as the default crypto implementation.
Crypto interface now is composed of 3 interfaces.

- MessageCrypto (contains message wrapping unwrapping and validation)
- BackendCrypto (contains Key manipulation, storage and conversion)
- CryptoUtils (utilities for tests)

Sealing the message (poW and Nonce) has been completely removed.

We keep the whisper structure of encrypted message with the format:
`flags+payloadSize+payload+padding+signature`
but only inside the default crypto implementation. Now the payload of the whisper message is a simple `[]byte`.

Padding is now an internal thing of the crypto implementation. Currently we have kept whisper padding, but maybe we can remove it.

Moved `crypto.go` to new package `pss/crypto`.

Benchmark tests have improved a bit mainly in larger keys:

This branch:
```bash
12:17 $ go test -cpu 4 -run=^$ -bench=^*$
goos: linux
goarch: amd64
pkg: github.com/ethersphere/swarm/pss
BenchmarkSymkeySend/256-4                           5000            270974 ns/op
BenchmarkSymkeySend/1024-4                          5000            281318 ns/op
BenchmarkSymkeySend/1048576-4                        100          12807973 ns/op
BenchmarkSymkeySend/10485760-4                        10         130643575 ns/op
BenchmarkSymkeySend/104857600-4                        2         637537803 ns/op
BenchmarkAsymkeySend/256-4                          2000           1016503 ns/op
BenchmarkAsymkeySend/1024-4                         2000            801877 ns/op
BenchmarkAsymkeySend/1048576-4                       100          19147333 ns/op
BenchmarkAsymkeySend/10485760-4                       10         208490108 ns/op
BenchmarkAsymkeySend/104857600-4                       1        1429648152 ns/op
PASS
ok      github.com/ethersphere/swarm/pss        18.953s
```

Master:
```bash
09:45 $ go test -cpu 4 -run=^$ -bench=^*$
goos: linux
goarch: amd64
pkg: github.com/ethersphere/swarm/pss
BenchmarkSymkeySend/256-4                           5000            280375 ns/op
BenchmarkSymkeySend/1024-4                          5000            289331 ns/op
BenchmarkSymkeySend/1048576-4                         50          23784979 ns/op
BenchmarkSymkeySend/10485760-4                         2         658122235 ns/op
BenchmarkSymkeySend/104857600-4                        1        4254846571 ns/op
BenchmarkAsymkeySend/256-4                          2000            818449 ns/op
BenchmarkAsymkeySend/1024-4                         2000            891359 ns/op
BenchmarkAsymkeySend/1048576-4                       100          31572557 ns/op
BenchmarkAsymkeySend/10485760-4                        2         521175499 ns/op
BenchmarkAsymkeySend/104857600-4                       1       33035929420 ns/op
PASS
ok      github.com/ethersphere/swarm/pss        54.628s
```