@jpeletier 2019-09-06
After this is merged, please ping me so I rebase #1734, as it builds on this work.