@kortatu 2019-09-09
> I am a bit skeptical whether this is not premature abstraction. I got the feeling that if it comes to actually use other crypto, it may not fit the interfaces designed here.
> 

About the "early abstraction", there is a second and third step in #1655 to define a crypto API ( Define a modular crypto API based on a selection of reference message crypto implementations). In factr @nolash talked about deciding the API in a SWIP. 
This "distilling" only pretends to show which features should contain a crypto API. Saying that, I'm sure this first step could be improved, so I will try to implement all the comments and hints in this PR.